Compute the cross-section loadings of yields of a canonical A0_N model

Compute the cross-section loadings of yields of a canonical A0_N model

Usage

A0N__BnAn(mat, K1XQ, ModelType, dX = NULL, r0 = NULL, SSX = NULL, Economies)

Arguments

mat

vector of maturities (J x 1). Maturities are in multiples of the discrete interval used in the model

K1XQ

risk-neutral feedback matrix (N x N)

ModelType

string-vector containing the label of the model to be estimated

dX

state loadings for the one-period rate (1xN). Default is a vector of ones

r0

the long run risk neutral mean of the short rate (scalar)

SSX

the covariance matrix of the errors (N x N)

Economies

string-vector containing the names of the economies which are part of the economic system

Value

List containing:

Intercept (Jx1)

slope (JxN)

the betan (JX1, part of the intercepts unrelated to the long run risk neutral mean r0) coefficients of a canonical A_0(N) model.

References

This function is a modified version of the "A0N__computeBnAn" function by Le and Singleton (2018).
"A Small Package of Matlab Routines for the Estimation of Some Term Structure Models."
(Euro Area Business Cycle Network Training School - Term Structure Modelling). Available at: https://cepr.org/40029

Dai and Singleton (2000). "Specification Analysis of Affine Term Structure Models" (The Journal of Finance)