Data: domestic risk factors - Candelon and Moura (2024, JFEC)

Domestic risk factors data used in Candelon and Moura (2024, JFEC)

Usage

data("DomMacro")

Format

A matrix of country-specific risk factors (inflation and economic activity growth) for Brazil, China, Mexico, and Uruguay. The data have monthly frequency and span the period from June/2004 to January/2020.

Source

Inflation

year-over-year variation from Consumer price index published by the International Monetary Fund <https://data.imf.org/en/datasets/IMF.STA:CPI>

Eonomic activity

GDP leading indicator published by the OECD <https://www.oecd.org/en/data/indicators/composite-leading-indicator-cli.html>

References

Candelon, B. and Moura, R. (2024) "A Multicountry Model of the Term Structures of Interest Rates with a GVAR". (Journal of Financial Econometrics)