Create the variable labels used in the estimation

Create the variable labels used in the estimation

Usage

ParaLabelsOpt(ModelType, WishStationarityQ, MLEinputs, BS_outputs = FALSE)

Arguments

ModelType

a string-vector containing the label of the model to be estimated

WishStationarityQ

User must set "1" is she wishes to impose the largest eigenvalue under the Q to be strictly smaller than 1. Otherwise set "0"

MLEinputs

Set of inputs that are necessary to the log-likelihood function

BS_outputs

Generates simplified output list in the bootstrap setting. Default is set to FALSE.

Value

list containing starting values and constraints: for each input argument K (of f), we need four inputs that look like:

a starting value: K0

a variable label ('K0') followed by a ':' followed by a type of constraint. The constraint can be:

'bounded': bounded matrix;

'Jordan' or 'Jordan MultiCountry': a matrix of Jordan type;

'psd': psd matrix;

'stationary': largest eigenvalue of the risk-neutral feedback matrix is strictly smaller than 1;

'diag' or 'BlockDiag': a diagonal or block diagonal matrix.

'JLLstructure': to impose the zero-restrictions on the variance-voriance matrix along the lines of the JLL models

a lower bound lb (lb <- NULL -> no lower bound)

an upper bound ub (ub <- NULL -> no upper bound)

Specification of the optimization settings:

'iter off': hide the printouts of the numerical optimization routines;

'fminunc only': only uses fminunc for the optimization;

”fminsearch only': only uses fminsearch for the optimization.