converts the vectorized auxiliary parameter vector x to the parameters that go directly into the likelihood function.

converts the vectorized auxiliary parameter vector x to the parameters that go directly into the likelihood function.

Usage

Update_ParaList(
  x,
  sizex,
  con,
  FactorLabels,
  Economies,
  JLLinputs = NULL,
  GVARinputs = NULL,
  ListInputSet
)

Arguments

x

vector containing all the vectorized auxiliary parameters

sizex

matrix (6x2) containing the size information of all parameters

con

if con = 'concentration', then set the value of the parameter whose name contains @ to empty

FactorLabels

string-list based which contains the labels of all the variables present in the model

Economies

string-vector containing the names of the economies which are part of the economic system

JLLinputs

Set of necessary inputs used in the estimation of the JLL-based models

GVARinputs

Set of necessary inputs used in the estimation of the GVAR-based models

ListInputSet

variable inputs used in the optimization (see "Optimization" function)

Value

same form as the list of parameters, except now the parameters are updated with the values provided by the auxiliary x. Importantly, by construction, all the constraints on the underlying parameters are satisfied.

References

This function is a modified version of the "update_para" function by Le and Singleton (2018).
"A Small Package of Matlab Routines for the Estimation of Some Term Structure Models."
(Euro Area Business Cycle Network Training School - Term Structure Modelling). Available at: https://cepr.org/40029